refactor(templates): isolate rpc functions
