+++ src/NodeAsyncHttpRuntime/LoadFileChunkLoadingRuntimeModule.js
@@ -11,7 +11,7 @@ const { getInitialChunkIds } = require("webpack/lib/javascript/StartupHelpers");
 const compileBooleanMatcher = require("webpack/lib/util/compileBooleanMatcher");
 const { getUndoPath } = require("webpack/lib/util/identifier");
 
-const rpcLoadTemplate = require("../templates/rpcLoad");
+const rpcLoadFromBaseURITemplate = require("../templates/rpcLoadFromBaseURI");
 
 class ReadFileChunkLoadingRuntimeModule extends RuntimeModule {
   constructor(runtimeRequirements, options, context) {
@@ -154,8 +154,8 @@ class ReadFileChunkLoadingRuntimeModule extends RuntimeModule {
                           )} + ${
                             RuntimeGlobals.getChunkScriptFilename
                           }(chunkId));`,
-                          rpcLoadTemplate,
-                          `rpcLoad(${
+                          rpcLoadFromBaseURITemplate,
+                          `rpcLoadFromBaseURI(${
                             getBaseUri
                               ? `await ${getBaseUri(baseURI)}`
                               : `"${baseURI}"`
@@ -201,8 +201,8 @@ class ReadFileChunkLoadingRuntimeModule extends RuntimeModule {
                 `var filename = require('path').join(__dirname, ${JSON.stringify(
                   rootOutputDir
                 )} + ${RuntimeGlobals.getChunkUpdateScriptFilename}(chunkId));`,
-                rpcLoadTemplate,
-                `rpcLoad(${
+                rpcLoadFromBaseURITemplate,
+                `rpcLoadFromBaseURI(${
                   promiseBaseURI ? `await ${promiseBaseURI}` : `"${baseURI}"`
                 }, ${
                   RuntimeGlobals.getChunkUpdateScriptFilename
